Stage 1 refactor genesis state and state applier into store
