+++ core/store/Cargo.toml
@@ -25,6 +25,7 @@ lru.workspace = true
 num_cpus.workspace = true
 once_cell.workspace = true
 rand.workspace = true
+rayon.workspace = true
 rlimit.workspace = true
 rocksdb.workspace = true
 serde.workspace = true
@@ -36,6 +37,7 @@ thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 
+near-chain-configs = { workspace = true, features = ["metrics"] }
 near-crypto.workspace = true
 near-fmt.workspace = true
 near-o11y.workspace = true
