+++ core/store/src/genesis_state_applier.rs
@@ -1,28 +1,27 @@
-use crate::config::RuntimeConfig;
-use crate::Runtime;
+use crate::flat::FlatStateChanges;
+use crate::{
+    get_account, get_received_data, set, set_access_key, set_account, set_code, set_delay_receipt,
+    set_postponed_receipt, set_received_data, ShardTries, TrieUpdate,
+};
 use borsh::BorshSerialize;
 use near_chain_configs::Genesis;
 use near_crypto::PublicKey;
 use near_primitives::account::{AccessKey, Account};
 use near_primitives::contract::ContractCode;
 use near_primitives::receipt::{DelayedReceiptIndices, Receipt, ReceiptEnum, ReceivedData};
+use near_primitives::runtime::config::RuntimeConfig;
 use near_primitives::runtime::fees::StorageUsageConfig;
 use near_primitives::shard_layout::ShardUId;
 use near_primitives::state_record::{state_record_to_account_id, StateRecord};
 use near_primitives::trie_key::TrieKey;
 use near_primitives::types::{AccountId, Balance, ShardId, StateChangeCause, StateRoot};
-use near_store::flat::FlatStateChanges;
-use near_store::{
-    get_account, get_received_data, set, set_access_key, set_account, set_code,
-    set_postponed_receipt, set_received_data, ShardTries, TrieUpdate,
-};
 use std::collections::{HashMap, HashSet};
 use std::sync::atomic;
 
 /// Computes the expected storage per account for a given stream of StateRecord(s).
 /// For example: the storage for Contract depends on its length, we don't charge storage for receipts
 /// and we compute a fixed (config-configured) number of bytes for each account (to store account id).
-pub(crate) struct StorageComputer<'a> {
+pub struct StorageComputer<'a> {
     /// Map from account id to number of storage bytes used.
     result: HashMap<AccountId, u64>,
     /// Configuration that keeps information like 'how many bytes should accountId consume' etc.
@@ -30,12 +29,12 @@ pub(crate) struct StorageComputer<'a> {
 }
 
 impl<'a> StorageComputer<'a> {
-    pub(crate) fn new(config: &'a RuntimeConfig) -> Self {
+    pub fn new(config: &'a RuntimeConfig) -> Self {
         Self { result: HashMap::new(), config: &config.fees.storage_usage_config }
     }
 
     /// Updates user's storage info based on the StateRecord.
-    pub(crate) fn process_record(&mut self, record: &StateRecord) {
+    pub fn process_record(&mut self, record: &StateRecord) {
         // Note: It's okay to use unsafe math here, because this method should only be called on the trusted
         // state records (e.g. at launch from genesis)
         let account_and_storage = match record {
@@ -68,14 +67,14 @@ impl<'a> StorageComputer<'a> {
     }
 
     /// Adds multiple StateRecords to the users' storage info.
-    pub(crate) fn process_records(&mut self, records: &[StateRecord]) {
+    pub fn process_records(&mut self, records: &[StateRecord]) {
         for record in records {
             self.process_record(record);
         }
     }
 
     /// Returns the current storage use for each user.
-    pub(crate) fn finalize(self) -> HashMap<AccountId, u64> {
+    pub fn finalize(self) -> HashMap<AccountId, u64> {
         self.result
     }
 }
@@ -243,8 +242,7 @@ impl GenesisStateApplier {
                     })
                 }
                 StateRecord::DelayedReceipt(receipt) => storage.modify(|state_update| {
-                    Runtime::delay_receipt(state_update, delayed_receipts_indices, &*receipt)
-                        .unwrap();
+                    set_delay_receipt(state_update, delayed_receipts_indices, &*receipt);
                 }),
             }
         });
