+++ core/store/src/lib.rs
@@ -16,6 +16,8 @@ pub use db::{
     CHUNK_TAIL_KEY, COLD_HEAD_KEY, FINAL_HEAD_KEY, FORK_TAIL_KEY, HEADER_HEAD_KEY, HEAD_KEY,
     LARGEST_TARGET_HEIGHT_KEY, LATEST_KNOWN_KEY, TAIL_KEY,
 };
+pub use genesis_state_applier::GenesisStateApplier;
+pub use genesis_state_applier::StorageComputer;
 use near_crypto::PublicKey;
 use near_fmt::{AbbrBytes, StorageKey};
 use near_primitives::account::{AccessKey, Account};
@@ -46,6 +48,8 @@ mod columns;
 pub mod config;
 pub mod db;
 pub mod flat;
+pub mod genesis;
+mod genesis_state_applier;
 pub mod metadata;
 pub mod metrics;
 pub mod migrations;
@@ -767,6 +771,23 @@ pub fn get_delayed_receipt_indices(
     trie: &dyn TrieAccess,
 ) -> Result<DelayedReceiptIndices, StorageError> {
     Ok(get(trie, &TrieKey::DelayedReceiptIndices)?.unwrap_or_default())
+}
+
+// Adds the given receipt into the end of the delayed receipt queue in the state.
+pub fn set_delay_receipt(
+    state_update: &mut TrieUpdate,
+    delayed_receipts_indices: &mut DelayedReceiptIndices,
+    receipt: &Receipt,
+) {
+    set(
+        state_update,
+        TrieKey::DelayedReceipt { index: delayed_receipts_indices.next_available_index },
+        receipt,
+    );
+    delayed_receipts_indices.next_available_index = delayed_receipts_indices
+        .next_available_index
+        .checked_add(1)
+        .expect("Next available index for delayed receipt exceeded the integer limit");
 }
 
 pub fn set_access_key(
